[test & fix & chore](定点数版本.目标跟踪.卡尔曼滤波): 添加测试案例，修复数学运算错误
- [fix]修复卡尔曼滤波Bug
    Source/fixed_point/Source/track_kalman.cc 缺少时间戳复制
    Source/fixed_point/Source/track_model.cc 雅可比矩阵计算错误

- [feat]卡尔曼滤波更新函数添加参数`bool need_pred_meas`，开关预测测量值的功能
    Source/fixed_point/Source/track_kalman.cc
    当卡尔曼滤波和数据关联结合时，更新这一步骤拆分为预测测量值和滤波。

- [test]添加测试案例，测试卡尔曼滤波和Eigen性质
    Test/Eigen/CMakeLists.txt
    Test/Eigen/test_eigen_assignment.cc
    Test/Eigen/test_eigen_size.cc
    Test/Radar_Fixed/test_track_kalman.cc
    Test/Radar_Fixed/test_track_kalman.data

- [fix] 修复函数参数使用传值而不是引用的问题
    Source/fixed_point/Include/track_deleter.hh
    Source/fixed_point/Include/track_initiator.hh
    Source/fixed_point/Source/track_deleter.cc
    Source/fixed_point/Source/track_initiator.cc

- [chore]clangformat格式化对齐数组
    .clang-format

- [chore]gitignore屏蔽Doc/Doxyfile
    .gitignore
    Doc/Doxyfile

+++ Source/fixed_point/Include/radar_math_types.h
@@ -20,7 +20,7 @@
 #include <stdint.h>
 #include <stddef.h>
 
-typedef float rd_float_t;
+typedef double rd_float_t;
 
 
 #include "radar_matrix.h"
